moving pose segmentation
